move utils back into lib
