+++ components/ui/form.tsx
@@ -10,7 +10,7 @@ import {
   useFormContext,
 } from "react-hook-form";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/lib/cn";
 import { Label } from "@/components/ui/label";
 
 const Form = FormProvider;
