+++ lib/generate-data-page.tsx
