+++ lib/generate-new-cropped-image-file.ts
