+++ lib/process-image.ts
